restcalc/requirements_docker.txt
cryptography==39.0.1
pytest
wheel
setuptools
Flask
SQLAlchemy
Werkzeug==2.3.2
markupsafe==2.1.1
sqlalchemy
python-dotenv
psycopg2-binary
autopep8
Flask-JWT-Extended
pyopenssl
marshmallow
bcrypt
flask-talisman
requests
pytest
flask_cors
boto3
redis